@@ -2,18 +2,25 @@
 
 
 namespace Onetech\WebAuthn\Attestation\Format;
+
 use Onetech\WebAuthn\Attestation\AuthenticatorData;
 use Onetech\WebAuthn\WebAuthnException;
 use Onetech\WebAuthn\Binary\ByteBuffer;
+use stdClass;
 
-class AndroidSafetyNet extends FormatBase {
-    private $_signature;
-    private $_signedValue;
-    private $_x5c;
-    private $_payload;
+class AndroidSafetyNet extends FormatBase
+{
+    private string $_signature;
+    private string $_signedValue;
+    private string|false $_x5c;
+    private stdClass $_payload;
 
-    public function __construct($AttestionObject, AuthenticatorData $authenticatorData) {
-        parent::__construct($AttestionObject, $authenticatorData);
+    /**
+     * @throws WebAuthnException
+     */
+    public function __construct($AttentionObject, AuthenticatorData $authenticatorData)
+    {
+        parent::__construct($AttentionObject, $authenticatorData);
 
         // check data
         $attStmt = $this->_attestationObject['attStmt'];
@@ -45,10 +52,10 @@ class AndroidSafetyNet extends FormatBase {
         $header = \json_decode($header);
         $payload = \json_decode($payload);
 
-        if (!($header instanceof \stdClass)) {
+        if (!($header instanceof stdClass)) {
             throw new WebAuthnException('invalid JWS header', WebAuthnException::INVALID_DATA);
         }
-        if (!($payload instanceof \stdClass)) {
+        if (!($payload instanceof stdClass)) {
             throw new WebAuthnException('invalid JWS payload', WebAuthnException::INVALID_DATA);
         }
 
@@ -65,7 +72,7 @@ class AndroidSafetyNet extends FormatBase {
         $this->_payload = $payload;
 
         if (count($header->x5c) > 1) {
-            for ($i=1; $i<count($header->x5c); $i++) {
+            for ($i = 1; $i < count($header->x5c); $i++) {
                 $this->_x5c_chain[] = \base64_decode($header->x5c[$i]);
             }
             unset ($i);
@@ -79,8 +86,9 @@ class AndroidSafetyNet extends FormatBase {
      * has been approved as a Google-certified Android device.
      * @return bool
      */
-    public function ctsProfileMatch() {
-        return isset($this->_payload->ctsProfileMatch) ? !!$this->_payload->ctsProfileMatch : false;
+    public function ctsProfileMatch(): bool
+    {
+        return isset($this->_payload->ctsProfileMatch) && !!$this->_payload->ctsProfileMatch;
     }
 
 
@@ -88,14 +96,18 @@ class AndroidSafetyNet extends FormatBase {
      * returns the key certificate in PEM format
      * @return string
      */
-    public function getCertificatePem() {
+    public function getCertificatePem(): ?string
+    {
         return $this->_createCertificatePem($this->_x5c);
     }
 
     /**
      * @param string $clientDataHash
+     * @return bool
+     * @throws WebAuthnException
      */
-    public function validateAttestation($clientDataHash) {
+    public function validateAttestation(string $clientDataHash): bool
+    {
         $publicKey = \openssl_pkey_get_public($this->getCertificatePem());
 
         // Verify that the nonce in the response is identical to the Base64 encoding
@@ -107,7 +119,7 @@ class AndroidSafetyNet extends FormatBase {
         // Verify that attestationCert is issued to the hostname "attest.android.com"
         $certInfo = \openssl_x509_parse($this->getCertificatePem());
         if (!\is_array($certInfo) || ($certInfo['subject']['CN'] ?? '') !== 'attest.android.com') {
-            throw new WebAuthnException('invalid certificate CN in JWS (' . ($certInfo['subject']['CN'] ?? '-'). ')', WebAuthnException::INVALID_DATA);
+            throw new WebAuthnException('invalid certificate CN in JWS (' . ($certInfo['subject']['CN'] ?? '-') . ')', WebAuthnException::INVALID_DATA);
         }
 
         // Verify that the basicIntegrity attribute in the payload of response is true.
@@ -126,7 +138,8 @@ class AndroidSafetyNet extends FormatBase {
      * @return boolean
      * @throws WebAuthnException
      */
-    public function validateRootCertificate($rootCas) {
+    public function validateRootCertificate(array $rootCas): bool
+    {
         $chainC = $this->_createX5cChainFile();
         if ($chainC) {
             $rootCas[] = $chainC;
@@ -145,7 +158,8 @@ class AndroidSafetyNet extends FormatBase {
      * @param string $data
      * @return string
      */
-    private function _base64url_decode($data) {
+    private function _base64url_decode(string $data): string
+    {
         return \base64_decode(\strtr($data, '-_', '+/') . \str_repeat('=', 3 - (3 + \strlen($data)) % 4));
     }
 }
