@@ -2,7 +2,9 @@
 
 
 namespace Onetech\WebAuthn\Binary;
+
 use Onetech\WebAuthn\WebAuthnException;
+use Random\RandomException;
 
 /**
  * Modified version of https://github.com/madwizard-thomas/webauthn-server/blob/master/src/Format/ByteBuffer.php
@@ -10,23 +12,25 @@ use Onetech\WebAuthn\WebAuthnException;
  * Modified by Lukas Buchs
  * Thanks Thomas for your work!
  */
-class ByteBuffer implements \JsonSerializable, \Serializable {
+class ByteBuffer implements \JsonSerializable, \Serializable
+{
     /**
      * @var bool
      */
-    public static $useBase64UrlEncoding = false;
+    public static bool $useBase64UrlEncoding = false;
 
     /**
      * @var string
      */
-    private $_data;
+    private string $_data;
 
     /**
      * @var int
      */
-    private $_length;
+    private int $_length;
 
-    public function __construct($binaryData) {
+    public function __construct($binaryData)
+    {
         $this->_data = (string)$binaryData;
         $this->_length = \strlen($binaryData);
     }
@@ -38,10 +42,12 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
 
     /**
      * create a ByteBuffer from a base64 url encoded string
-     * @param string $base64url
+     * @param $base64url
      * @return ByteBuffer
+     * @throws WebAuthnException
      */
-    public static function fromBase64Url($base64url): ByteBuffer {
+    public static function fromBase64Url($base64url): ByteBuffer
+    {
         $bin = self::_base64url_decode($base64url);
         if ($bin === false) {
             throw new WebAuthnException('ByteBuffer: Invalid base64 url string', WebAuthnException::BYTEBUFFER);
@@ -51,10 +57,12 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
 
     /**
      * create a ByteBuffer from a base64 url encoded string
-     * @param string $hex
+     * @param $hex
      * @return ByteBuffer
+     * @throws WebAuthnException
      */
-    public static function fromHex($hex): ByteBuffer {
+    public static function fromHex($hex): ByteBuffer
+    {
         $bin = \hex2bin($hex);
         if ($bin === false) {
             throw new WebAuthnException('ByteBuffer: Invalid hex string', WebAuthnException::BYTEBUFFER);
@@ -64,10 +72,13 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
 
     /**
      * create a random ByteBuffer
-     * @param string $length
+     * @param $length
      * @return ByteBuffer
+     * @throws WebAuthnException
+     * @throws RandomException
      */
-    public static function randomBuffer($length): ByteBuffer {
+    public static function randomBuffer($length): ByteBuffer
+    {
         if (\function_exists('random_bytes')) { // >PHP 7.0
             return new ByteBuffer(\random_bytes($length));
 
@@ -83,21 +94,33 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
     // PUBLIC
     // -----------------------
 
-    public function getBytes($offset, $length): string {
+    /**
+     * @throws WebAuthnException
+     */
+    public function getBytes($offset, $length): string
+    {
         if ($offset < 0 || $length < 0 || ($offset + $length > $this->_length)) {
             throw new WebAuthnException('ByteBuffer: Invalid offset or length', WebAuthnException::BYTEBUFFER);
         }
         return \substr($this->_data, $offset, $length);
     }
 
-    public function getByteVal($offset): int {
+    /**
+     * @throws WebAuthnException
+     */
+    public function getByteVal($offset): int
+    {
         if ($offset < 0 || $offset >= $this->_length) {
             throw new WebAuthnException('ByteBuffer: Invalid offset', WebAuthnException::BYTEBUFFER);
         }
         return \ord(\substr($this->_data, $offset, 1));
     }
 
-    public function getJson($jsonFlags=0) {
+    /**
+     * @throws WebAuthnException
+     */
+    public function getJson($jsonFlags = 0)
+    {
         $data = \json_decode($this->getBinaryString(), null, 512, $jsonFlags);
         if (\json_last_error() !== JSON_ERROR_NONE) {
             throw new WebAuthnException(\json_last_error_msg(), WebAuthnException::BYTEBUFFER);
@@ -105,18 +128,34 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
         return $data;
     }
 
-    public function getLength(): int {
+    /**
+     * @return int
+     */
+    public function getLength(): int
+    {
         return $this->_length;
     }
 
-    public function getUint16Val($offset) {
+    /**
+     * @param $offset
+     * @return mixed
+     * @throws WebAuthnException
+     */
+    public function getUint16Val($offset): mixed
+    {
         if ($offset < 0 || ($offset + 2) > $this->_length) {
             throw new WebAuthnException('ByteBuffer: Invalid offset', WebAuthnException::BYTEBUFFER);
         }
         return unpack('n', $this->_data, $offset)[1];
     }
 
-    public function getUint32Val($offset) {
+    /**
+     * @param $offset
+     * @return mixed
+     * @throws WebAuthnException
+     */
+    public function getUint32Val($offset): mixed
+    {
         if ($offset < 0 || ($offset + 4) > $this->_length) {
             throw new WebAuthnException('ByteBuffer: Invalid offset', WebAuthnException::BYTEBUFFER);
         }
@@ -129,7 +168,13 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
         return $val;
     }
 
-    public function getUint64Val($offset) {
+    /**
+     * @param $offset
+     * @return mixed
+     * @throws WebAuthnException
+     */
+    public function getUint64Val($offset): mixed
+    {
         if (PHP_INT_SIZE < 8) {
             throw new WebAuthnException('ByteBuffer: 64-bit values not supported by this system', WebAuthnException::BYTEBUFFER);
         }
@@ -146,7 +191,13 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
         return $val;
     }
 
-    public function getHalfFloatVal($offset) {
+    /**
+     * @param $offset
+     * @return float|int
+     * @throws WebAuthnException
+     */
+    public function getHalfFloatVal($offset)
+    {
         //FROM spec pseudo decode_half(unsigned char *halfp)
         $half = $this->getUint16Val($offset);
 
@@ -164,14 +215,22 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
         return ($half & 0x8000) ? -$val : $val;
     }
 
-    public function getFloatVal($offset) {
+    /**
+     * @throws WebAuthnException
+     */
+    public function getFloatVal($offset)
+    {
         if ($offset < 0 || ($offset + 4) > $this->_length) {
             throw new WebAuthnException('ByteBuffer: Invalid offset', WebAuthnException::BYTEBUFFER);
         }
         return unpack('G', $this->_data, $offset)[1];
     }
 
-    public function getDoubleVal($offset) {
+    /**
+     * @throws WebAuthnException
+     */
+    public function getDoubleVal($offset)
+    {
         if ($offset < 0 || ($offset + 8) > $this->_length) {
             throw new WebAuthnException('ByteBuffer: Invalid offset', WebAuthnException::BYTEBUFFER);
         }
@@ -181,7 +240,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
     /**
      * @return string
      */
-    public function getBinaryString(): string {
+    public function getBinaryString(): string
+    {
         return $this->_data;
     }
 
@@ -189,28 +249,28 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * @param string|ByteBuffer $buffer
      * @return bool
      */
-    public function equals($buffer): bool {
-        if (is_object($buffer) && $buffer instanceof ByteBuffer) {
+    public function equals(ByteBuffer|string $buffer): bool
+    {
+        if ($buffer instanceof ByteBuffer) {
             return $buffer->getBinaryString() === $this->getBinaryString();
-
-        } else if (is_string($buffer)) {
+        } else {
             return $buffer === $this->getBinaryString();
         }
-        
-        return false;
     }
 
     /**
      * @return string
      */
-    public function getHex(): string {
+    public function getHex(): string
+    {
         return \bin2hex($this->_data);
     }
 
     /**
      * @return bool
      */
-    public function isEmpty(): bool {
+    public function isEmpty(): bool
+    {
         return $this->_length === 0;
     }
 
@@ -220,7 +280,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * return binary data in RFC 1342-Like serialized string
      * @return string
      */
-    public function jsonSerialize(): string {
+    public function jsonSerialize(): string
+    {
         if (ByteBuffer::$useBase64UrlEncoding) {
             return self::_base64url_encode($this->_data);
 
@@ -233,7 +294,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * Serializable-Interface
      * @return string
      */
-    public function serialize(): string {
+    public function serialize(): string
+    {
         return \serialize($this->_data);
     }
 
@@ -241,7 +303,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * Serializable-Interface
      * @param string $serialized
      */
-    public function unserialize($serialized) {
+    public function unserialize($serialized)
+    {
         $this->_data = \unserialize($serialized);
         $this->_length = \strlen($this->_data);
     }
@@ -250,7 +313,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * (PHP 8 deprecates Serializable-Interface)
      * @return array
      */
-    public function __serialize(): array {
+    public function __serialize(): array
+    {
         return [
             'data' => \serialize($this->_data)
         ];
@@ -260,7 +324,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * object to string
      * @return string
      */
-    public function __toString(): string {
+    public function __toString(): string
+    {
         return $this->getHex();
     }
 
@@ -269,7 +334,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * @param array $data
      * @return void
      */
-    public function __unserialize($data) {
+    public function __unserialize(array $data)
+    {
         if ($data && isset($data['data'])) {
             $this->_data = \unserialize($data['data']);
             $this->_length = \strlen($this->_data);
@@ -285,7 +351,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * @param string $data
      * @return string
      */
-    protected static function _base64url_decode($data): string {
+    protected static function _base64url_decode(string $data): string
+    {
         return \base64_decode(\strtr($data, '-_', '+/') . \str_repeat('=', 3 - (3 + \strlen($data)) % 4));
     }
 
@@ -294,7 +361,8 @@ class ByteBuffer implements \JsonSerializable, \Serializable {
      * @param string $data
      * @return string
      */
-    protected static function _base64url_encode($data): string {
+    protected static function _base64url_encode(string $data): string
+    {
         return \rtrim(\strtr(\base64_encode($data), '+/', '-_'), '=');
     }
 }
