@@ -1,58 +1,60 @@
 <?php
 
 namespace Onetech\WebAuthn\Attestation;
+
 use Onetech\WebAuthn\WebAuthnException;
 use Onetech\WebAuthn\CBOR\CborDecoder;
 use Onetech\WebAuthn\Binary\ByteBuffer;
+use stdClass;
+use function strlen;
 
 /**
  * @author Lukas Buchs
  * @license https://github.com/lbuchs/WebAuthn/blob/master/LICENSE MIT
  */
-class AuthenticatorData {
+class AuthenticatorData
+{
     protected $_binary;
     protected $_rpIdHash;
     protected $_flags;
     protected $_signCount;
     protected $_attestedCredentialData;
-    protected $_extensionData;
-
-
 
     // Cose encoded keys
-    private static $_COSE_KTY = 1;
-    private static $_COSE_ALG = 3;
+    private static int $_COSE_KTY = 1;
+    private static int $_COSE_ALG = 3;
 
     // Cose curve
-    private static $_COSE_CRV = -1;
-    private static $_COSE_X = -2;
-    private static $_COSE_Y = -3;
+    private static int $_COSE_CRV = -1;
+    private static int $_COSE_X = -2;
+    private static int $_COSE_Y = -3;
 
     // Cose RSA PS256
-    private static $_COSE_N = -1;
-    private static $_COSE_E = -2;
+    private static int $_COSE_N = -1;
+    private static int $_COSE_E = -2;
 
     // EC2 key type
-    private static $_EC2_TYPE = 2;
-    private static $_EC2_ES256 = -7;
-    private static $_EC2_P256 = 1;
+    private static int $_EC2_TYPE = 2;
+    private static int $_EC2_ES256 = -7;
+    private static int $_EC2_P256 = 1;
 
     // RSA key type
-    private static $_RSA_TYPE = 3;
-    private static $_RSA_RS256 = -257;
+    private static int $_RSA_TYPE = 3;
+    private static int $_RSA_RS256 = -257;
 
     // OKP key type
-    private static $_OKP_TYPE = 1;
-    private static $_OKP_ED25519 = 6;
-    private static $_OKP_EDDSA = -8;
+    private static int $_OKP_TYPE = 1;
+    private static int $_OKP_ED25519 = 6;
+    private static int $_OKP_EDDSA = -8;
 
     /**
      * Parsing the authenticatorData binary.
      * @param string $binary
      * @throws WebAuthnException
      */
-    public function __construct($binary) {
-        if (!\is_string($binary) || \strlen($binary) < 37) {
+    public function __construct(string $binary)
+    {
+        if (strlen($binary) < 37) {
             throw new WebAuthnException('Invalid authenticatorData input', WebAuthnException::INVALID_DATA);
         }
         $this->_binary = $binary;
@@ -90,8 +92,9 @@ class AuthenticatorData {
      * @return string
      * @throws WebAuthnException
      */
-    public function getAAGUID() {
-        if (!($this->_attestedCredentialData instanceof \stdClass)) {
+    public function getAAGUID(): string
+    {
+        if (!($this->_attestedCredentialData instanceof stdClass)) {
             throw  new WebAuthnException('credential data not included in authenticator data', WebAuthnException::INVALID_DATA);
         }
         return $this->_attestedCredentialData->aaguid;
@@ -101,7 +104,8 @@ class AuthenticatorData {
      * returns the authenticatorData as binary
      * @return string
      */
-    public function getBinary() {
+    public function getBinary(): string
+    {
         return $this->_binary;
     }
 
@@ -110,8 +114,9 @@ class AuthenticatorData {
      * @return string
      * @throws WebAuthnException
      */
-    public function getCredentialId() {
-        if (!($this->_attestedCredentialData instanceof \stdClass)) {
+    public function getCredentialId(): string
+    {
+        if (!($this->_attestedCredentialData instanceof stdClass)) {
             throw  new WebAuthnException('credential id not included in authenticator data', WebAuthnException::INVALID_DATA);
         }
         return $this->_attestedCredentialData->credentialId;
@@ -120,18 +125,27 @@ class AuthenticatorData {
     /**
      * returns the public key in PEM format
      * @return string
+     * @throws WebAuthnException
      */
-    public function getPublicKeyPem() {
-        if (!($this->_attestedCredentialData instanceof \stdClass) || !isset($this->_attestedCredentialData->credentialPublicKey)) {
+    public function getPublicKeyPem(): string
+    {
+        if (!($this->_attestedCredentialData instanceof stdClass) || !isset($this->_attestedCredentialData->credentialPublicKey)) {
             throw  new WebAuthnException('credential data not included in authenticator data', WebAuthnException::INVALID_DATA);
         }
-        
+
         $der = null;
         switch ($this->_attestedCredentialData->credentialPublicKey->kty ?? null) {
-            case self::$_EC2_TYPE: $der = $this->_getEc2Der(); break;
-            case self::$_RSA_TYPE: $der = $this->_getRsaDer(); break;
-            case self::$_OKP_TYPE: $der = $this->_getOkpDer(); break;
-            default: throw new WebAuthnException('invalid key type', WebAuthnException::INVALID_DATA);
+            case self::$_EC2_TYPE:
+                $der = $this->_getEc2Der();
+                break;
+            case self::$_RSA_TYPE:
+                $der = $this->_getRsaDer();
+                break;
+            case self::$_OKP_TYPE:
+                $der = $this->_getOkpDer();
+                break;
+            default:
+                throw new WebAuthnException('invalid key type', WebAuthnException::INVALID_DATA);
         }
 
         $pem = '-----BEGIN PUBLIC KEY-----' . "\n";
@@ -145,23 +159,25 @@ class AuthenticatorData {
      * @return string
      * @throws WebAuthnException
      */
-    public function getPublicKeyU2F() {
-        if (!($this->_attestedCredentialData instanceof \stdClass) || !isset($this->_attestedCredentialData->credentialPublicKey)) {
+    public function getPublicKeyU2F(): string
+    {
+        if (!($this->_attestedCredentialData instanceof stdClass) || !isset($this->_attestedCredentialData->credentialPublicKey)) {
             throw  new WebAuthnException('credential data not included in authenticator data', WebAuthnException::INVALID_DATA);
         }
         if (($this->_attestedCredentialData->credentialPublicKey->kty ?? null) !== self::$_EC2_TYPE) {
             throw new WebAuthnException('signature algorithm not ES256', WebAuthnException::INVALID_PUBLIC_KEY);
         }
         return "\x04" . // ECC uncompressed
-                $this->_attestedCredentialData->credentialPublicKey->x .
-                $this->_attestedCredentialData->credentialPublicKey->y;
+            $this->_attestedCredentialData->credentialPublicKey->x .
+            $this->_attestedCredentialData->credentialPublicKey->y;
     }
 
     /**
-     * returns the SHA256 hash of the relying party id (=hostname)
+     * returns the SHA256 hash of the relying on party id (=hostname)
      * @return string
      */
-    public function getRpIdHash() {
+    public function getRpIdHash(): string
+    {
         return $this->_rpIdHash;
     }
 
@@ -169,7 +185,8 @@ class AuthenticatorData {
      * returns the sign counter
      * @return int
      */
-    public function getSignCount() {
+    public function getSignCount(): int
+    {
         return $this->_signCount;
     }
 
@@ -177,7 +194,8 @@ class AuthenticatorData {
      * returns true if the user is present
      * @return boolean
      */
-    public function getUserPresent() {
+    public function getUserPresent(): bool
+    {
         return $this->_flags->userPresent;
     }
 
@@ -185,7 +203,8 @@ class AuthenticatorData {
      * returns true if the user is verified
      * @return boolean
      */
-    public function getUserVerified() {
+    public function getUserVerified(): bool
+    {
         return $this->_flags->userVerified;
     }
 
@@ -193,7 +212,7 @@ class AuthenticatorData {
      * returns true if the backup is eligible
      * @return boolean
      */
-    public function getIsBackupEligible()
+    public function getIsBackupEligible(): bool
     {
         return $this->_flags->isBackupEligible;
     }
@@ -202,7 +221,7 @@ class AuthenticatorData {
      * returns true if the current credential is backed up
      * @return boolean
      */
-    public function getIsBackup()
+    public function getIsBackup(): bool
     {
         return $this->_flags->isBackup;
     }
@@ -214,8 +233,10 @@ class AuthenticatorData {
     /**
      * Returns DER encoded EC2 key
      * @return string
+     * @throws WebAuthnException
      */
-    private function _getEc2Der() {
+    private function _getEc2Der(): string
+    {
         return $this->_der_sequence(
             $this->_der_sequence(
                 $this->_der_oid("\x2A\x86\x48\xCE\x3D\x02\x01") . // OID 1.2.840.10045.2.1 ecPublicKey
@@ -229,7 +250,8 @@ class AuthenticatorData {
      * Returns DER encoded EdDSA key
      * @return string
      */
-    private function _getOkpDer() {
+    private function _getOkpDer(): string
+    {
         return $this->_der_sequence(
             $this->_der_sequence(
                 $this->_der_oid("\x2B\x65\x70") // OID 1.3.101.112 curveEd25519 (EdDSA 25519 signature algorithm)
@@ -242,7 +264,8 @@ class AuthenticatorData {
      * Returns DER encoded RSA key
      * @return string
      */
-    private function _getRsaDer() {
+    private function _getRsaDer(): string
+    {
         return $this->_der_sequence(
             $this->_der_sequence(
                 $this->_der_oid("\x2A\x86\x48\x86\xF7\x0D\x01\x01\x01") . // OID 1.2.840.113549.1.1.1 rsaEncryption
@@ -260,10 +283,11 @@ class AuthenticatorData {
     /**
      * reads the flags from flag byte
      * @param string $binFlag
-     * @return \stdClass
+     * @return stdClass
      */
-    private function _readFlags($binFlag) {
-        $flags = new \stdClass();
+    private function _readFlags(string $binFlag): stdClass
+    {
+        $flags = new stdClass();
 
         $flags->bit_0 = !!($binFlag & 1);
         $flags->bit_1 = !!($binFlag & 2);
@@ -288,12 +312,13 @@ class AuthenticatorData {
      * read attested data
      * @param string $binary
      * @param int $endOffset
-     * @return \stdClass
+     * @return stdClass
      * @throws WebAuthnException
      */
-    private function _readAttestData($binary, &$endOffset) {
-        $attestedCData = new \stdClass();
-        if (\strlen($binary) <= 55) {
+    private function _readAttestData(string $binary, int &$endOffset): stdClass
+    {
+        $attestedCData = new stdClass();
+        if (strlen($binary) <= 55) {
             throw new WebAuthnException('Attested data should be present but is missing', WebAuthnException::INVALID_DATA);
         }
 
@@ -316,36 +341,45 @@ class AuthenticatorData {
     /**
      * reads COSE key-encoded elliptic curve public key in EC2 format
      * @param string $binary
+     * @param $offset
      * @param int $endOffset
-     * @return \stdClass
+     * @return stdClass
      * @throws WebAuthnException
      */
-    private function _readCredentialPublicKey($binary, $offset, &$endOffset) {
+    private function _readCredentialPublicKey(string $binary, $offset, int &$endOffset): stdClass
+    {
         $enc = CborDecoder::decodeInPlace($binary, $offset, $endOffset);
 
         // COSE key-encoded elliptic curve public key in EC2 format
-        $credPKey = new \stdClass();
+        $credPKey = new stdClass();
         $credPKey->kty = $enc[self::$_COSE_KTY];
         $credPKey->alg = $enc[self::$_COSE_ALG];
 
         switch ($credPKey->alg) {
-            case self::$_EC2_ES256: $this->_readCredentialPublicKeyES256($credPKey, $enc); break;
-            case self::$_RSA_RS256: $this->_readCredentialPublicKeyRS256($credPKey, $enc); break;
-            case self::$_OKP_EDDSA: $this->_readCredentialPublicKeyEDDSA($credPKey, $enc); break;
+            case self::$_EC2_ES256:
+                $this->_readCredentialPublicKeyES256($credPKey, $enc);
+                break;
+            case self::$_RSA_RS256:
+                $this->_readCredentialPublicKeyRS256($credPKey, $enc);
+                break;
+            case self::$_OKP_EDDSA:
+                $this->_readCredentialPublicKeyEDDSA($credPKey, $enc);
+                break;
         }
 
         return $credPKey;
     }
 
     /**
-     * extract EDDSA informations from cose
-     * @param \stdClass $credPKey
-     * @param \stdClass $enc
+     * extract EDDSA information's from cose
+     * @param stdClass $credPKey
+     * @param stdClass $enc
      * @throws WebAuthnException
      */
-    private function _readCredentialPublicKeyEDDSA(&$credPKey, $enc) {
+    private function _readCredentialPublicKeyEDDSA(stdClass &$credPKey, stdClass $enc): void
+    {
         $credPKey->crv = $enc[self::$_COSE_CRV];
-        $credPKey->x   = $enc[self::$_COSE_X] instanceof ByteBuffer ? $enc[self::$_COSE_X]->getBinaryString() : null;
+        $credPKey->x = $enc[self::$_COSE_X] instanceof ByteBuffer ? $enc[self::$_COSE_X]->getBinaryString() : null;
         unset ($enc);
 
         // Validation
@@ -361,21 +395,22 @@ class AuthenticatorData {
             throw new WebAuthnException('curve not Ed25519', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
-        if (\strlen($credPKey->x) !== 32) {
+        if (strlen($credPKey->x) !== 32) {
             throw new WebAuthnException('Invalid X-coordinate', WebAuthnException::INVALID_PUBLIC_KEY);
         }
     }
 
     /**
-     * extract ES256 informations from cose
-     * @param \stdClass $credPKey
-     * @param \stdClass $enc
+     * extract ES256 information's from cose
+     * @param stdClass $credPKey
+     * @param stdClass $enc
      * @throws WebAuthnException
      */
-    private function _readCredentialPublicKeyES256(&$credPKey, $enc) {
+    private function _readCredentialPublicKeyES256(stdClass &$credPKey, stdClass $enc): void
+    {
         $credPKey->crv = $enc[self::$_COSE_CRV];
-        $credPKey->x   = $enc[self::$_COSE_X] instanceof ByteBuffer ? $enc[self::$_COSE_X]->getBinaryString() : null;
-        $credPKey->y   = $enc[self::$_COSE_Y] instanceof ByteBuffer ? $enc[self::$_COSE_Y]->getBinaryString() : null;
+        $credPKey->x = $enc[self::$_COSE_X] instanceof ByteBuffer ? $enc[self::$_COSE_X]->getBinaryString() : null;
+        $credPKey->y = $enc[self::$_COSE_Y] instanceof ByteBuffer ? $enc[self::$_COSE_Y]->getBinaryString() : null;
         unset ($enc);
 
         // Validation
@@ -391,22 +426,23 @@ class AuthenticatorData {
             throw new WebAuthnException('curve not P-256', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
-        if (\strlen($credPKey->x) !== 32) {
+        if (strlen($credPKey->x) !== 32) {
             throw new WebAuthnException('Invalid X-coordinate', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
-        if (\strlen($credPKey->y) !== 32) {
+        if (strlen($credPKey->y) !== 32) {
             throw new WebAuthnException('Invalid Y-coordinate', WebAuthnException::INVALID_PUBLIC_KEY);
         }
     }
 
     /**
-     * extract RS256 informations from COSE
-     * @param \stdClass $credPKey
-     * @param \stdClass $enc
+     * extract RS256 information's from COSE
+     * @param stdClass $credPKey
+     * @param stdClass $enc
      * @throws WebAuthnException
      */
-    private function _readCredentialPublicKeyRS256(&$credPKey, $enc) {
+    private function _readCredentialPublicKeyRS256(stdClass &$credPKey, stdClass $enc): void
+    {
         $credPKey->n = $enc[self::$_COSE_N] instanceof ByteBuffer ? $enc[self::$_COSE_N]->getBinaryString() : null;
         $credPKey->e = $enc[self::$_COSE_E] instanceof ByteBuffer ? $enc[self::$_COSE_E]->getBinaryString() : null;
         unset ($enc);
@@ -420,11 +456,11 @@ class AuthenticatorData {
             throw new WebAuthnException('signature algorithm not ES256', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
-        if (\strlen($credPKey->n) !== 256) {
+        if (strlen($credPKey->n) !== 256) {
             throw new WebAuthnException('Invalid RSA modulus', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
-        if (\strlen($credPKey->e) !== 3) {
+        if (strlen($credPKey->e) !== 3) {
             throw new WebAuthnException('Invalid RSA public exponent', WebAuthnException::INVALID_PUBLIC_KEY);
         }
 
@@ -433,16 +469,15 @@ class AuthenticatorData {
     /**
      * reads cbor encoded extension data.
      * @param string $binary
-     * @return array
+     * @return void
      * @throws WebAuthnException
      */
-    private function _readExtensionData($binary) {
+    private function _readExtensionData(string $binary): void
+    {
         $ext = CborDecoder::decode($binary);
         if (!\is_array($ext)) {
             throw new WebAuthnException('invalid extension data', WebAuthnException::INVALID_DATA);
         }
-
-        return $ext;
     }
 
 
@@ -450,7 +485,8 @@ class AuthenticatorData {
     // DER functions
     // ---------------
 
-    private function _der_length($len) {
+    private function _der_length($len): string
+    {
         if ($len < 128) {
             return \chr($len);
         }
@@ -459,27 +495,32 @@ class AuthenticatorData {
             $lenBytes = \chr($len % 256) . $lenBytes;
             $len = \intdiv($len, 256);
         }
-        return \chr(0x80 | \strlen($lenBytes)) . $lenBytes;
+        return \chr(0x80 | strlen($lenBytes)) . $lenBytes;
     }
 
-    private function _der_sequence($contents) {
-        return "\x30" . $this->_der_length(\strlen($contents)) . $contents;
+    private function _der_sequence($contents): string
+    {
+        return "\x30" . $this->_der_length(strlen($contents)) . $contents;
     }
 
-    private function _der_oid($encoded) {
-        return "\x06" . $this->_der_length(\strlen($encoded)) . $encoded;
+    private function _der_oid($encoded): string
+    {
+        return "\x06" . $this->_der_length(strlen($encoded)) . $encoded;
     }
 
-    private function _der_bitString($bytes) {
-        return "\x03" . $this->_der_length(\strlen($bytes) + 1) . "\x00" . $bytes;
+    private function _der_bitString($bytes): string
+    {
+        return "\x03" . $this->_der_length(strlen($bytes) + 1) . "\x00" . $bytes;
     }
 
-    private function _der_nullValue() {
+    private function _der_nullValue(): string
+    {
         return "\x05\x00";
     }
 
-    private function _der_unsignedInteger($bytes) {
-        $len = \strlen($bytes);
+    private function _der_unsignedInteger($bytes): string
+    {
+        $len = strlen($bytes);
 
         // Remove leading zero bytes
         for ($i = 0; $i < ($len - 1); $i++) {
@@ -496,6 +537,6 @@ class AuthenticatorData {
             $bytes = "\x00" . $bytes;
         }
 
-        return "\x02" . $this->_der_length(\strlen($bytes)) . $bytes;
+        return "\x02" . $this->_der_length(strlen($bytes)) . $bytes;
     }
 }
